requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
amqp==5.0.6
apiclient==1.0.4
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
billiard==3.6.4.0
black==21.7b0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-celery-beat==2.2.1
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-filter==2.4.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
flower==1.0.0
gevent==21.8.0
greenlet==1.1.0
grequests==0.6.0
humanize==3.11.0
idna==3.2
kombu==5.1.0
Markdown==3.3.4
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pycares==4.0.0
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
redis==3.5.3
regex==2021.8.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tomli==1.2.1
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0
